Add files via upload
CRUD OPERATION : This crud operation is built by using spring boot in this two tables are there one is employee table and department table , for security i'll implement JWT Implementation and for documentation i'll implement swagger implementation in this CRUD operation there is api colled insertuser , updateUser , deleteUser , getUser api are there. all this api will working properly we can test this api by using postman with the help of url ,we write a json data in the body of the tomcat and then we apply a respective method ie. post , put , get etc ..